extracting ebutt1object base into generic ebuttdocumentbase class
