+++ ebu_tt_live/documents/__init__.py
@@ -1,7 +1,6 @@
-from .base import SubtitleDocument, TimeBase, DocumentSequence
-from .ebutt1 import EBUTT1Document, EBUTT1ObjectBase
-from .ebutt3 import EBUTT3Document, EBUTT3DocumentSequence, EBUTTAuthorsGroupControlRequest, EBUTT3ObjectBase, \
-    EBUTTLiveMessage
+from .base import SubtitleDocument, TimeBase, DocumentSequence, EBUTTDocumentBase
+from .ebutt1 import EBUTT1Document
+from .ebutt3 import EBUTT3Document, EBUTT3DocumentSequence, EBUTTAuthorsGroupControlRequest, EBUTTLiveMessage
 from .ebuttd import EBUTTDDocument
 from .converters import ebutt3_to_ebuttd, EBUTT3EBUTTDConverter
