+++ ebu_tt_live/documents/base.py
@@ -1,5 +1,7 @@
 
 from ebu_tt_live.utils import ComparableMixin
+from ebu_tt_live import bindings
+from ebu_tt_live.bindings import _ebuttlm as ebuttlm
 
 # NOTE: Some of the code below includes handling of SMPTE time base, which was removed from version 1.0 of the specification.
 
@@ -82,4 +84,26 @@ class CloningDocumentSequence(DocumentSequence):
         :param kwargs: parameter override to constructor
         :return:
         """
+        raise NotImplementedError()
+
+
+class EBUTTDocumentBase(object):
+    message_type_mapping = {}
+
+    def get_xml(self):
+        raise NotImplementedError()
+
+    def get_dom(self):
+        raise NotImplementedError()
+
+    @classmethod
+    def create_from_xml(cls, xml):
+        instance = bindings.CreateFromDocument(
+            xml_text=xml
+        )
+        if isinstance(instance, ebuttlm.message_type):
+            return cls.message_type_mapping[instance.header.type].create_from_raw_binding(instance)
+
+    @classmethod
+    def create_from_raw_binding(cls, **kwargs):
         raise NotImplementedError()
